necessary files created and theme and manifest file modified for accessing camera
