+++ app/src/main/res/layout/activity_register.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="16dp"
-    tools:context=".RegisterActivity">
+    tools:context=".login_register.RegisterActivity">
 
     <TextView
         android:text="@string/register"
