+++ app/src/main/res/values/strings.xml
@@ -7,4 +7,6 @@
     <string name="new_user">New user? Register</string>
     <string name="confirm_password">Confirm Password</string>
     <string name="already_registered">Already registered? Login</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
